Always move bytes between the detour location and the patch address
This fixes sub-block patching. Previously, if we patched a sub basic
block region, patcherex would insert a detour at the beginning of the
basic block and only copy the instructions that fill the size of the
detour jump instruction and then at the end of the trampoline, it
expects to be able to jump back to that point. This is the incorrect
behavior since if we patch a sub basic block region, only the first
instruction would get copied over, and the remaining instructions in
between the detour location and the address we want to insert our patch
would not be executed.